requirements.txt
pillow
rpi.gpio